Refactor namespace from wayedges to mjolnir in way edges processing files
